% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{angle}
\alias{angle}
\title{Angle between two subspaces}
\usage{
angle(B1, B2)
}
\arguments{
\item{B1}{first matrix}

\item{B2}{second matrix}
}
\description{
Measures angle between two subspaces
 from http://www4.stat.ncsu.edu/~li/software/GroupDR.R
}
\examples{
b1 <- matrix(rnorm(10 * 2), ncol = 2)
b2 <- matrix(rnorm(10 * 2), ncol = 2)
angle(b1, b2)
}
