% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orthnorm}
\alias{orthnorm}
\title{Gram-Schmidt orthonormalization}
\usage{
orthnorm(X)
}
\arguments{
\item{X}{matrix to orthonormalize}
}
\description{
Gram-Schmidt orthonormalization
}
\examples{
X <- matrix(rnorm(10 * 2), ncol = 2)
X
orthnorm(X)
}
