% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiparametric_sdr.R
\name{phd}
\alias{phd}
\title{PHD SDR fitting function}
\usage{
phd(x, y, d = 5L)
}
\arguments{
\item{x}{an n x p matrix of covariates, where each row is an observation and each column is a predictor}

\item{y}{vector of responses of length n}

\item{d}{an integer representing the structural dimension}
}
\description{
fits SDR models (PHD approach)
}
