% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_sphd.R
\name{plot.hier_sdr_fit}
\alias{plot.hier_sdr_fit}
\title{Plotting hierarchical SDR models}
\usage{
\method{plot}{hier_sdr_fit}(x, ...)
}
\arguments{
\item{x}{fitted object returned by \code{\link[hierSDR]{hier.sphd}}}

\item{...}{not used}
}
\value{
No return value, called for side effects
}
\description{
Plots hier.sdr objects
}
\examples{

library(hierSDR)

}
\seealso{
\code{\link[hierSDR]{hier.sphd}} for function which fits hierarchical SDR model
}
