% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj.norm}
\alias{proj.norm}
\title{Norm of difference of projections}
\usage{
proj.norm(B1, B2)
}
\arguments{
\item{B1}{first matrix}

\item{B1}{second matrix}
}
\description{
Measures distance between two subspaces
}
\examples{
b1 <- matrix(rnorm(10 * 2), ncol = 2)
b2 <- matrix(rnorm(10 * 2), ncol = 2)
proj.norm(b1, b2)

}
